Make user change grid size with prompt
Implemented the feature that allows the user
to enter a number which is then stored in the
"size" variable. After the current grid is
eliminated, it creates another one by calling the
cloning functions and passing the "size" variable
as the argument. The "rows" nodeList is now
declared directly in the "cloneSquare" function,
otherwise the function would not work as it
wouldn't know how many rows have been
previously cloned with "cloneRow"